feat:  create utility module for emoji picker
https://youtrack.weseek.co.jp/issue/GW-7652
- Create helper for emoji picker
- Adjust implementation of emoji picker component
